requirements.txt
Django==3.1.0
djangorestframework==3.12.2
django-cors-headers==3.5.0